Extract otlp-parser module from otlp-shell
Mirror the pprof-parser layout for the OTLP profiling format:
- io.jafar.otlp.api: OtlpProfile model and OtlpParser entry point
- io.jafar.otlp.internal: protobuf wire-format decoder
- MinimalOtlpBuilder test fixture and the reader round-trip test move
  along; otlp-shell and jfr-mcp consume the parser API
