+++ jfr-mcp/src/test/java/io/jafar/mcp/OtlpHandlerIntegrationTest.java
@@ -2,7 +2,7 @@ package io.jafar.mcp;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import io.jafar.otlp.shell.MinimalOtlpBuilder;
+import io.jafar.otlp.MinimalOtlpBuilder;
 import io.modelcontextprotocol.server.McpSyncServerExchange;
 import io.modelcontextprotocol.spec.McpSchema.CallToolResult;
 import java.lang.reflect.Method;
