+++ otlp-parser/build.gradle
@@ -1,0 +1,50 @@
+plugins {
+    id 'java-library'
+    id 'java-test-fixtures'
+    id 'maven-publish'
+    id 'com.vanniktech.maven.publish'
+}
+
+def libraryName = "otlp-parser"
+def component_version = project.hasProperty("jafar_version") ? project.jafar_version : rootProject.version
+
+repositories {
+    mavenCentral()
+    mavenLocal()
+}
+
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(21)
+    }
+}
+
+dependencies {
+    // Shared protobuf wire-format decoding (ProtoUtil)
+    implementation project(':parser-core')
+
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.10.2'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.10.2'
+    testRuntimeOnly 'org.junit.platform:junit-platform-launcher:1.10.2'
+}
+
+test {
+    useJUnitPlatform()
+}
+
+javadoc {
+    options.addStringOption('Xdoclint:none', '-quiet')
+    failOnError = false
+}
+
+// Keep test fixtures (MinimalOtlpBuilder) out of the published artifact
+components.java.withVariantsFromConfiguration(configurations.testFixturesApiElements) { skip() }
+components.java.withVariantsFromConfiguration(configurations.testFixturesRuntimeElements) { skip() }
+
+group = 'io.btrace'
+version = component_version
+description = 'Fast OpenTelemetry Profiling (OTLP) parser with minimal API'
+
+mavenPublishing {
+    coordinates('io.btrace', 'jafar-otlp-parser', version.toString())
+}
