+++ otlp-parser/src/main/java/io/jafar/otlp/api/OtlpProfile.java
@@ -1,4 +1,4 @@
-package io.jafar.otlp.shell;
+package io.jafar.otlp.api;
 
 import java.util.List;
