+++ otlp-shell/build.gradle
@@ -1,6 +1,5 @@
 plugins {
     id 'java-library'
-    id 'java-test-fixtures'
 }
 
 def libraryName = "otlp-shell"
@@ -21,12 +20,13 @@ dependencies {
     // Shell infrastructure
     api project(':shell-core')
 
-    // Shared protobuf wire-format decoding (ProtoUtil)
-    implementation project(':parser-core')
+    // OTLP profile parser
+    api project(':otlp-parser')
 
     // Logging
     implementation 'org.slf4j:slf4j-api:2.0.5'
 
+    testImplementation(testFixtures(project(':otlp-parser')))
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.10.2'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.10.2'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher:1.10.2'
