+++ settings.gradle
@@ -42,6 +42,7 @@ include ':hdump-parser'
 include ':hdump-shell'
 include ':pprof-parser'
 include ':pprof-shell'
+include ':otlp-parser'
 include ':otlp-shell'
 include ':jafar-shell'
 include ':jfr2pprof'
